changed ptr to use uintptr_t

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jrinta- <jrinta-@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/07 19:07:43 by jrinta-           #+#    #+#             */
-/*   Updated: 2024/11/11 14:23:00 by jrinta-          ###   ########.fr       */
+/*   Updated: 2024/11/13 14:48:18 by jrinta-          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,13 +40,13 @@ int		is_spec(char c);
 int		is_flag(char c);
 int		print_hex_handler(unsigned int nbr, int capital, t_flags *flags);
 int		print_char(char c, t_flags flags);
-int		print_ptr_handler(unsigned long int n, t_flags flags);
+int		print_ptr_handler(uintptr_t n, t_flags flags);
 int		print_unsigned_handler(unsigned int nbr, t_flags flags);
 int		print_nbr_handler(int nbr, t_flags *flags);
 int		print_str_handler(const char *str, t_flags flags);
 int		print_s(const char *str);
 int		ft_printf(const char *, ...);
 char	*ft_ltoa(long n);
-char	*ft_xtoa(unsigned long int n, int capital);
+char	*ft_xtoa(uintptr_t n, int capital);
 
 #endif
